requirements.txt
streamlit 
streamlit-drawable-canvas 
pillow 
requests
python-dotenv